+++ src/app/Http/Controllers/ProductController.php
@@ -17,9 +17,11 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+        $id = Auth::User()->id;
+
         DB::table('products')->insert([
             'product_type_id' => $request->product_type_id,
-            'user_id' => $request->user_id,
+            'user_id' => $id,
             'product_name' => $request->product_name,
             'description' => $request->description,
             'product_picture' => $request->product_picture,
@@ -34,7 +36,7 @@ class ProductController extends Controller
 
     public function index()
     {
-        return Product::all();
+        //return Product::all();
     }
 
     public function create(Request $request)
@@ -56,33 +58,46 @@ class ProductController extends Controller
     public function destroy($id)
     {
         $product = Product::find($id);
-        $product->delete();
+        
+        if($product->user_id == Auth::User()->id){
+            $product->delete();
 
-        return "Product Deleted";
+            return "Product Deleted";
+        }else{
+            return "Cannot Delete";
+        }
     }
 
     public function update(Request $request, $id)
     {
-        $product = Product::find($id);
+        if($product->user_id == Auth::User()->id){
+            $product = Product::find($id);
 
-        $product->product_type_id = $request->product_type_id;
-        $product->user_id = $request->user_id;
-        $product->product_name = $request->product_name;
-        $product->price = $request->price;
-        $product->units = $request->units;
-        $product->description = $request->description;
-        $product->product_picture = $request->product_picture;
-        $product->updated_at = date('Y-m-d H:i:s');
-        $product->save();
-
-        return "Product Updated";
+            $product->product_type_id = $request->product_type_id;
+            $product->user_id = $request->user_id;
+            $product->product_name = $request->product_name;
+            $product->price = $request->price;
+            $product->units = $request->units;
+            $product->description = $request->description;
+            $product->product_picture = $request->product_picture;
+            $product->updated_at = date('Y-m-d H:i:s');
+            $product->save();
+    
+            return "Product Updated";
+        }else{
+            return "Cannot Update";
+        }
     }
 
     public function show($id)
     {
         $product = Product::find($id);
 
-        return response()->json($product);
+        if($product->user_id == Auth::User()->id){
+            return response()->json($product);
+        }else{
+            return "Cannot Show";
+        }
     }
 
     public function edit($id)
@@ -93,7 +108,8 @@ class ProductController extends Controller
     public function search(Request $request)
     {
         $search = $request->search;
-        $id = $request->id;
+        $id = Auth::User()->id;
+        //$id = $request->id;
         //$search = $request->getContent();
         //$search = explode("search=", $search);
         //$product = DB::table('products')
@@ -107,7 +123,8 @@ class ProductController extends Controller
 
     public function searchByUserID(Request $request)
     {
-        $id = $request->id;
+        //$id = $request->id;
+        $id = Auth::User()->id;
         $product = Product::where('user_id', $id)->get();
 
         return response()->json($product, 200);
@@ -115,7 +132,8 @@ class ProductController extends Controller
 
     public function productStock(Request $request)
     {
-        $id = $request->id;
+        //$id = $request->id;
+        $id = Auth::User()->id;
         //$product = Product::where('user_id', $id)->get();
 
         $product = DB::table('products')
@@ -136,7 +154,7 @@ class ProductController extends Controller
 
     public function productAuth()
     {
-        $data = "Welcome " . Auth::User()->username;
+        $data = "Welcome " . Auth::User()->id;
         return response()->json($data, 200);
     }
